requirements.txt
# Core FastAPI dependencies
fastapi[standard]==0.116.1
pytest==8.4.1

# PostgreSQL and database dependencies 
asyncpg==0.29.0
psycopg2-binary==2.9.10
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0

# Redis caching
redis[hiredis]==5.2.1
aioredis==2.0.1

# JWT Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.18

# Configuration and logging 
python-dotenv==1.0.1
structlog==24.4.0
pydantic-settings==2.6.1

# Security and middleware
slowapi==0.1.9
prometheus-client==0.21.1

# WebAuthn/Passkey dependencies
cbor2==5.6.5

# Testing dependencies
pytest-asyncio==0.24.0
pytest-docker==3.1.1
httpx==0.28.1
pytest-cov==6.0.0
aiosqlite==0.21.0